GWT bugfix
if number of graphs on a view change (e.g. due to loading of a different game or changing geometry), the previous graphs must be removed from the page/parent before new ones get added (i.e. clearing the graph list is not enough). this currently applies to MVPop2D/3D, MVMean, MVHistogram.